+++ NovaChat.Server/Controllers/GroupManagementController.cs
@@ -28,22 +28,22 @@ public class GroupManagementController : ControllerBase
     {
         if (!TryGetCurrentUserId(out var actorId)) return Unauthorized();
 
-        var chat = await _db.Chats.FirstOrDefaultAsync(c => c.Id == chatId && c.Type == ChatType.Group);
+        var chat = await _db.Chats.FirstOrDefaultAsync(c => c.Id == chatId && c.Type == (int)ChatType.Group);
         if (chat == null) return NotFound(new { message = "Group not found." });
 
         var actor = await _db.ChatMembers.FirstOrDefaultAsync(m => m.ChatId == chatId && m.UserId == actorId);
         var target = await _db.ChatMembers.FirstOrDefaultAsync(m => m.ChatId == chatId && m.UserId == userId);
-        if (actor == null || actor.Role != ChatMemberRole.Owner) return Forbid();
+        if (actor == null || actor.Role != (int)ChatMemberRole.Owner) return Forbid();
         if (target == null) return NotFound(new { message = "Member not found." });
-        if (target.Role == ChatMemberRole.Owner) return BadRequest(new { message = "The group owner cannot be changed to another role." });
+        if (target.Role == (int)ChatMemberRole.Owner) return BadRequest(new { message = "The group owner cannot be changed to another role." });
 
-        target.Role = request?.IsAdmin == true ? ChatMemberRole.Admin : ChatMemberRole.Member;
+        target.Role = request?.IsAdmin == true ? (int)ChatMemberRole.Admin : (int)ChatMemberRole.Member;
         await _db.SaveChangesAsync();
 
         var recipients = await _db.ChatMembers.Where(m => m.ChatId == chatId).Select(m => m.UserId.ToString()).ToListAsync();
         await _hub.Clients.Users(recipients).SendAsync("GroupUpdated", new { Id = chat.Id, Type = "Group", Name = chat.Name, AvatarUrl = chat.AvatarUrl, CreatedByUserId = chat.CreatedByUserId?.ToString() ?? string.Empty });
 
-        return Ok(new { message = target.Role == ChatMemberRole.Admin ? "Member promoted to admin." : "Admin privileges removed." });
+        return Ok(new { message = target.Role == (int)ChatMemberRole.Admin ? "Member promoted to admin." : "Admin privileges removed." });
     }
 
     [HttpPost("{chatId:int}/leave")]
@@ -51,12 +51,12 @@ public class GroupManagementController : ControllerBase
     {
         if (!TryGetCurrentUserId(out var userId)) return Unauthorized();
 
-        var chat = await _db.Chats.AsNoTracking().FirstOrDefaultAsync(c => c.Id == chatId && c.Type == ChatType.Group);
+        var chat = await _db.Chats.AsNoTracking().FirstOrDefaultAsync(c => c.Id == chatId && c.Type == (int)ChatType.Group);
         if (chat == null) return NotFound(new { message = "Group not found." });
 
         var member = await _db.ChatMembers.FirstOrDefaultAsync(m => m.ChatId == chatId && m.UserId == userId);
         if (member == null) return Forbid();
-        if (member.Role == ChatMemberRole.Owner) return BadRequest(new { message = "The owner cannot leave the group. Delete the group from Group Info instead." });
+        if (member.Role == (int)ChatMemberRole.Owner) return BadRequest(new { message = "The owner cannot leave the group. Delete the group from Group Info instead." });
 
         _db.ChatMembers.Remove(member);
         await _db.SaveChangesAsync();
@@ -72,10 +72,10 @@ public class GroupManagementController : ControllerBase
     {
         if (!TryGetCurrentUserId(out var userId)) return Unauthorized();
 
-        var chat = await _db.Chats.FirstOrDefaultAsync(c => c.Id == chatId && c.Type == ChatType.Group);
+        var chat = await _db.Chats.FirstOrDefaultAsync(c => c.Id == chatId && c.Type == (int)ChatType.Group);
         if (chat == null) return NotFound(new { message = "Group not found." });
 
-        var owner = await _db.ChatMembers.FirstOrDefaultAsync(m => m.ChatId == chatId && m.UserId == userId && m.Role == ChatMemberRole.Owner);
+        var owner = await _db.ChatMembers.FirstOrDefaultAsync(m => m.ChatId == chatId && m.UserId == userId && m.Role == (int)ChatMemberRole.Owner);
         if (owner == null) return Forbid();
 
         var recipients = await _db.ChatMembers.Where(m => m.ChatId == chatId).Select(m => m.UserId.ToString()).ToListAsync();
